Bump networkx version
The previous networkx uses a deprecated numpy feature
which causes an error with the following message:

```
`np.int` was a deprecated alias for the builtin `int`.
To avoid this error in existing code, use `int` by itself.
Doing this will not modify any behavior and is safe.
When replacing `np.int`, you may wish to use e.g. `np.int64` or `np.int32` to specify the precision.
If you wish to review your current use, check the release note link for additional information.
The aliases was originally deprecated in NumPy 1.20; for more details and guidance see the original release note at:
https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations. Did you mean: 'inf'?
```

This change happend with numpy>=v1.20.0. Networkx package fixes its own
dependency with `networkx>=3.0`.